refactor(type_checker): extract check() into stmt.rs
